+++ CHANGES.md
@@ -1,3 +1,8 @@
+unreleased
+----------
+
+- Support `ADD` new `--unpack` flag. (@MisterDA, #241)
+
 v8.3.1 2025-09-15
 -----------------